refactor: move server to separate module
