+++ package.json
@@ -7,7 +7,7 @@
     "start:dev": "nodemon",
     "start:prod": "webpack && node ./dist/main.js",
     "lint:fix": "eslint --fix --cache",
-    "test": "jest --verbose"
+    "test": "jest --verbose --maxWorkers=1"
   },
   "repository": {
     "type": "git",
